% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{corrAnalysis}
\alias{corrAnalysis}
\title{Perform gene correlation analysis on the gene expression matrix data, using
method indicated by users}
\usage{
corrAnalysis(
  geneCounts = NULL,
  method = "pearson",
  filePath = NULL,
  save = FALSE
)
}
\arguments{
\item{geneCounts}{A data frame containing the matrix of gene expression data,
normally obtained from RNA-seq or other sequencing experiments.
The row names of the matrix should be the genes, and the column names
should be the sample IDs. Default value is NULL.}

\item{method}{A character string indicating which correlation coefficient is
to be computed. One of "pearson" (default), "kendall", or "spearman". If
you are not sure which coefficient to use, just use "pearson".
\href{https://www.R-project.org/}{Link for more detail about correlation
coefficient}}

\item{filePath}{A character string path to the directory that the user want
to store the correlation analysis result. It is recommended to create a
new directory to store the output .csv file. Default value is NULL.
Should in the format: "/Path/to/the/directory".}

\item{save}{A boolean value to indicate if store the result into a file
at the file path. If set to TRUE, the output would be stored at the
filePath. Default value is FALSE.}
}
\value{
A correlation coefficient matrix that has n * n dimensions (n is the
number of genes). Each cell contains the correlation coefficient between
corresponding genes. The function would save the matrix into a .csv file.
}
\description{
This is a function that can conduct genetic correlation analysis to identify
pairs of genes that have a relationship in their expression. It calculates
the correlation coefficients between each pair of genes and store them into a
matrix. Different ways for the calculation of the correlation coefficients
are provided.
}
\examples{
# Example 1:
# Using gene counts matrix (geneCountsCorrelation) available with the package
\dontrun{
dim(geneCountsCorrelation) # 105 rows, 41 columns

# Perform gene correlation analysis on the input gene counts matrix available
# with the package
# This creates a correlation_analysis_result.csv file stores the result
# in the current working directory
corrResult <- corrAnalysis(geneCountsCorrelation,
                           method = "pearson")
corrResult[1:5, 1:5]
}

# Example 2:
# Use a different correlation coefficient
\dontrun{
corrResult2 <- corrAnalysis(geneCountsCorrelation,
                           method = "kendall")
corrResult2[1:5, 1:5]
}

# Example 3:
# Use a different correlation coefficient
\dontrun{
corrResult3 <- corrAnalysis(geneCountsCorrelation,
                           method = "spearman")
corrResult3[1:5, 1:5]
}

}
\references{
Geistlinger L, Csaba G, Zimmer R (2016). “Bioconductor's EnrichmentBrowser:
seamless navigation through combined results of set- & network-based
enrichment analysis.” BMC Bioinformatics, 17, 45.
doi:10.1186/s12859-016-0884-1.

R Core Team (2023). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria. https://www.R-project.org/
}
