% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffExpressionResult}
\alias{diffExpressionResult}
\title{The output of the differential expression analysis, used as the input
for function "extractSignificantGene" and "labelGenes"}
\format{
A matrix with 18309 rows and 6 columns, row names are gene names:
\describe{
\item{baseMean}{The mean of normalized counts of all samples, normalizing
for sequencing depth}
\item{log2FoldChange}{The effect size estimate, which tells about how much
the gene's expression seems to have changed in
different samples}
\item{lfcSE}{Standard error of the log2FoldChange.}
\item{stat}{log2FoldChange divided by lfcSE}
\item{pvalue}{Use stat to compare to a standard Normal distribution to
generate a two-tailed pvalue, which tells about whether the
difference is significant}
\item{padj}{Adjusted p value}
}
}
\source{
Raw data obtained from Gene Expression Omnibus, generated by
GEAnaly::diffExpressionAnalysis.
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157103}{link}
}
\usage{
diffExpressionResult
}
\description{
Generated by the differential expression analysis function based on DESeq2.
The data is a matrix that describes how different the expression levels of
genes are in different samples.
}
\examples{
\dontrun{
 diffExpressionResult
}
}
\keyword{datasets}
