% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichAnalysis.R
\name{enrichAnalysis}
\alias{enrichAnalysis}
\title{Perform genes' functional enrichment analysis based on the functions provided
by gprofiler2, which is the procedure of identifying functions that are over-
or under-represented among a set of genes.}
\usage{
enrichAnalysis(
  significantGenes = NULL,
  pvalueCutoff = 0.05,
  correctionMethod = "g_SCS",
  filePath = NULL,
  save = FALSE
)
}
\arguments{
\item{significantGenes}{A data frame, using the genes names as row names of
the data frame. It does not matter what other contents are. Make sure that
row.names(significantGenes) is the vector that contains the genes for the
enrichment analysis.}

\item{pvalueCutoff}{A positive double, which is the threshold set by the
users used to decide if the enrichment of a pathway is significant.
Only pathways with smaller p-value are tagged as significant. Default
value is 0.05. The pvalueCutoff should be a value between 0 and 1.}

\item{correctionMethod}{A character string indicating the algorithm used for
multiple testing correction, one of "g_SCS" (default), "fdr", "bonferroni".
If you are not sure choose which one, just use g_SCS}

\item{filePath}{A character string path to the directory that the user want
to store the enrichment analysis result. It is recommended to create a
new directory to store the output .tsv file. Default value is NULL.
Should in the format: "/Path/to/the/directory".}

\item{save}{A boolean value to indicate if store the result into a file
at the file path. If set to TRUE, the output would be stored at the
filePath. Default value is FALSE.}
}
\value{
A named list containing the named list of results from
gprofiler2 and a processed data frame based on the results from
gprofiler2 for visualization.
\itemize{
\item gProfilerResult - A named list where 'result' contains a data frame
with the enrichment analysis results and 'meta' contains metadata needed
for visualization.
\item enrichmentVis - A processed data frame, which is a subset of the
original enrichment analysis results, and also the geneRatio of different
pathways are calculated and added for visualizations.
}
}
\description{
This is a function that can conduct functional enrichment analysis to identify
functions that are over- or under-represented among the genes inputted. The
function is based on the "gprofiler2" package. The function uses pathways
databases including GO, KEGG, REAC, TF, MIRNA, CORUM, HP, HPA, WP as data
sources for the enrichment analysis.
}
\examples{
# Example 1:
# Use the significant gene list available with the package
\dontrun{
dim(significantGenes) # 1299 rows, 6 columns

# Perform enrichment analysis
enrichOutputListE <- enrichAnalysis(significantGenes,
                                   pvalueCutoff = 0.05,
                                   correctionMethod = "g_SCS")

# Access the results stored in the list returned
enrichOutputListE$gProfilerResult
enrichOutputListE$enrichmentVis
}

# Example 2:
# P value can be adjusted and correction method can also be changed
\dontrun{
# Perform enrichment analysis
enrichOutputListE2 <- enrichAnalysis(significantGenes,
                                   pvalueCutoff = 0.1,
                                   correctionMethod = "fdr")

# Access the results stored in the list returned
enrichOutputListE2$gProfilerResult
enrichOutputListE2$enrichmentVis
}

}
\references{
Geistlinger L, Csaba G, Zimmer R (2016). “Bioconductor's EnrichmentBrowser:
seamless navigation through combined results of set- & network-based
enrichment analysis.” BMC Bioinformatics, 17, 45. doi:10.1186/s12859-016-0884-1.

Kolberg L, Raudvere U, Kuzmin I, Vilo J, Peterson H (2020). “gprofiler2-an R package for
gene list functional enrichment analysis and namespace conversion toolset g:Profiler.”
F1000Research, 9 (ELIXIR)(709). R package version 0.2.2.

R Core Team (2023). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria. https://www.R-project.org/
}
