% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrichOutputList}
\alias{enrichOutputList}
\title{The output of the enrichment analysis, used as the input
for function "visEnrichAnaly" and "visEnrichAnalyLollipop"}
\format{
A named list of length 2, containing results of enrichment analysis:
\describe{
\item{gProfilerResult}{A named list where 'result' contains a data frame
with the enrichment analysis results and 'meta' contains metadata needed
for visualization.}
\item{enrichmentVis}{A processed data frame, which is a subset of the
original enrichment analysis results, and also the geneRatio of different
pathways are calculated and added for visualizations.}
}
}
\source{
Raw data obtained from Gene Expression Omnibus, generated by
GEAnaly::enrichAnalysis.
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157103}{link}
}
\usage{
enrichOutputList
}
\description{
A named list containing the named list of results from
gprofiler2 and a processed data frame based on the results from
gprofiler2 for visualization. Generated by the enrichment analysis function
based on gprofiler2.
}
\examples{
\dontrun{
 enrichOutputList
}
}
\keyword{datasets}
