% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDeGene.R
\name{extractSignificantGene}
\alias{extractSignificantGene}
\title{Filter out the genes that have significantly different expression levels by
the user-customized threshold}
\usage{
extractSignificantGene(
  diffExpressionResult = NULL,
  filePath = NULL,
  pValue = 0.1,
  foldChange = 1
)
}
\arguments{
\item{diffExpressionResult}{A DESeqResults object, a simple subclass of
DataFrame defined by DESeq2. It contains the columns: baseMean,
log2FoldChange, lfcSE, stat, pvalue and padj, and also includes
metadata columns of variable information. This input is the output
of function diffExpressionAnalysis. Default value is NULL.}

\item{filePath}{A character string path to the directory that the user want
to store the significant genes selected by the p value and fold change
threshold. It is recommended to create a new directory to store the output
.csv file. Default value is NULL. Should in the format:
"/Path/to/the/directory/".}

\item{pValue}{A positive double, which is the threshold set by the
users used to filter genes that are differentially expressed in different
samples. Only genes that has smaller p values than the threshold would be
regarded as significant genes. Default value is 0.1. pValue is used to
measure if a difference is significant. The pValue should be a value
between 0 and 1.}

\item{foldChange}{A positive integer, which is the threshold set by the users
used to filter genes that are differentially expressed in different
samples. Only genes that has larger absolute values of fold change than
the threshold would be regarded as genes express differently. Default
value is 1. Fold change is a measure describing how much the expression
levels of genes change between different samples.}
}
\value{
Returns a data frame, which is a subset of the input dataframe
"diffExpressionResult", containing the genes that have significantly
different expression level in different samples (based on the input
thresholds). It also contains the columns: baseMean, log2FoldChange,
lfcSE, stat, pvalue and padj, and also includes
metadata columns of variable information.
}
\description{
This is a function that select genes that have significantly different
expression levels in different samples according to the p value and
fold change provided by the users. As long as the gene has smaller p value
than the p value threshold and has larger fold change value than the
fold change threshold inputted by the users, these genes are regarded as
genes have significantly different expression level in different samples.
The function save these significant genes into a csv file.
}
\examples{
# Example 1:

}
\references{
Wickham H., François R., Henry L., Müller K., Vaughan D. (2023).
\emph{dplyr: A Grammar of Data Manipulation}. R package version 1.1.3,
\href{https://CRAN.R-project.org/package=dplyr}{link}.

R Core Team (2023). R: A Language and Environment for Statistical Computing. R Foundation
for Statistical Computing, Vienna, Austria. \href{https://www.R-project.org/}{link}.

Geistlinger L, Csaba G, Zimmer R (2016). “Bioconductor's EnrichmentBrowser:
seamless navigation through combined results of set- & network-based
enrichment analysis.” BMC Bioinformatics, 17, 45. doi:10.1186/s12859-016-0884-1.
}
