% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneCorResult}
\alias{geneCorResult}
\title{A correlation matrix recording the correlation coefficient between genes}
\format{
A matrix with 103 rows and 103 columns, row names and column names
are same gene names. Each cell records the correlation coefficient of
corresponding genes on the x and y axis
}
\source{
Raw data obtained from Gene Expression Omnibus, generated by
GEAnaly::corrAnalysis.
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157103}{link}
}
\usage{
geneCorResult
}
\description{
A data frame contains the correlation coefficient of each pair of genes,
values on the diagonal are all zero.
}
\examples{
\dontrun{
 geneCorResult
}
}
\keyword{datasets}
