% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneCountsCorrelation}
\alias{geneCountsCorrelation}
\title{The gene expression matrix data for genes in different samples}
\format{
A matrix with 1000 rows and 41 columns, row names are gene names,
column names are sample IDs:
\describe{
\item{Column names}{Columns starting with C are sample with Covid. Samples
starting with NC are samples without Covid}
}
}
\source{
Gene Expression Omnibus
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157103}{link}
}
\usage{
geneCountsCorrelation
}
\description{
A data frame containing the matrix of un-normalized read
counts of gene expression data, normally obtained from RNA-seq or other
sequencing experiments. The count numbers in the matrix are
non-negative integers. The row names of the matrix are the genes,
and the column names are the sample IDs. This is a subset of
geneCountsDiffExpression, selecting only the first 1000 genes and 41 samples,
since the original data is too large for the correlatin analysis, which takes
too long for an example.
}
\examples{
\dontrun{
 geneCountsCorrelation
}
}
\keyword{datasets}
