% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneCountsDiffExpression}
\alias{geneCountsDiffExpression}
\title{The gene expression matrix data for genes in different samples}
\format{
A matrix with 11000 rows and 126 columns, row names are gene names,
column names are sample IDs:
\describe{
\item{Column names}{Columns, whose name start with C, are samples with Covid.
Columns, whose name start with NC, are samples without Covid.}
}
}
\source{
Gene Expression Omnibus
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157103}{link}
}
\usage{
geneCountsDiffExpression
}
\description{
A data frame containing the matrix of un-normalized read
counts of gene expression data, normally obtained from RNA-seq or other
sequencing experiments. The count numbers in the matrix are
non-negative integers. The row names of the matrix are the genes,
and the column names are the sample IDs. Since the limit of the data size for
an R package, geneCountsDiffExpression is a subset of the original data.
geneCountsDiffExpression just uses the first 11000 genes.
}
\examples{
\dontrun{
 geneCountsDiffExpression
}
}
\keyword{datasets}
