% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{significantGenes}
\alias{significantGenes}
\title{A data frame that is the subset of the output of differential expression
analysis, containing genes that have significantly different expression
levels in different samples.}
\format{
A data frame with 1299 rows and 6 columns, row names are sample names:
\describe{
\item{baseMean}{The mean of normalized counts of all samples, normalizing
for sequencing depth}
\item{log2FoldChange}{The effect size estimate, which tells about how much
the gene's expression seems to have changed in
different samples}
\item{lfcSE}{Standard error of the log2FoldChange.}
\item{stat}{log2FoldChange divided by lfcSE}
\item{pvalue}{Use stat to compare to a standard Normal distribution to
generate a two-tailed pvalue, which tells about whether the
difference is significant}
\item{padj}{Adjusted p value}
}
}
\source{
Raw data obtained from Gene Expression Omnibus, generated by
GEAnaly::diffExpressionAnalysis and GEAnaly::extractSignificantGene.
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157103}{link}
}
\usage{
significantGenes
}
\description{
The data has the same columns as diffExpressionResult, but with fewer genes.
Genes in the data all have p values that are smaller than 0.05, and
absolute values of log2FoldChange that are larger than 2, which means that
these genes express significantly differently in different samples.
}
\examples{
\dontrun{
 significantGenes
}
}
\keyword{datasets}
