% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visCorrelationAnaly}
\alias{visCorrelationAnaly}
\title{Generate a heatmap for the gene correlation analysis result}
\usage{
visCorrelationAnaly(
  corMatrix = NULL,
  colors = c("steelblue", "white", "brown"),
  filePath = NULL,
  save = FALSE
)
}
\arguments{
\item{corMatrix}{A correlation coefficient matrix that has n * n dimensions
(n is the number of genes). Each cell contains the correlation coefficient
between corresponding genes. Output of function "enrichAnalysis" is a good
example.}

\item{colors}{A vector that has three colors used. Default as
c("steelblue","white","brown"). For the default value, negative values are
marked with relatively "steelblue" color, positive values are marked with
relatively "brown" color. 0 is marked as "white".}

\item{filePath}{A character string path to the directory that the user want
to store the visualization of the correlation analysis result.
It is recommended to create a new directory to store the output .png file.
Default value is NULL. Should in the format: "/Path/to/the/directory".}

\item{save}{A boolean value to indicate if store the result into a file
at the file path. If set to TRUE, the output would be stored at the
filePath. Default value is FALSE.}
}
\value{
A ggplot object, and save the output visualization plot to "filePath"
if save is TRUE
}
\description{
This is a function that can visualize the gene correlation analysis result
as a heatmap, whose x-axis and y-axis are gene names, each cell in the
heatmap indicates the correlation coefficient of corresponding genes.
}
\examples{
# Example 1:
# Using correlation matrix (geneCorResult) available with the package
\dontrun{
dim(geneCorResult) # 103 rows, 103 columns

# Plot a heatmap for the correlation analysis results
visCorrelationAnaly(geneCorResult)

}

}
\references{
Kolde R. (2019). pheatmap: Pretty Heatmaps. R package version 1.0.12,
https://CRAN.R-project.org/package=pheatmap

R Core Team (2023). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/

Wickham H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4
https://CRAN.R-project.org/package=ggplot2

Yu G. (2023). ggplotify: Convert Plot to 'grob' or 'ggplot' Object.
R package version 0.1.2, \url{https://CRAN.R-project.org/package=ggplotify}.
}
