% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visEnrichAnaly}
\alias{visEnrichAnaly}
\title{Generate a Manhattan plot for the enrichment analysis result}
\usage{
visEnrichAnaly(
  enrichOutputList = NULL,
  interactive = TRUE,
  filePath = NULL,
  save = FALSE
)
}
\arguments{
\item{enrichOutputList}{A named list containing the named list of enrichment
analysis results from gprofiler2 and a processed data frame based on the
results from gprofiler2 for visualization. The output of function
"enrichAnalysis" is a good example.}

\item{interactive}{A boolean indicating whether to output an interactive plot,
Default as TRUE.}

\item{filePath}{A character string path to the directory that the user want
to store the visualization of the enrichment analysis result.
It is recommended to create a new directory to store the output .png/.html
file. Default value is NULL. Should in the format: "/Path/to/the/directory".}

\item{save}{A boolean value to indicate if store the result into a file
at the file path. If set to TRUE, the output would be stored at the
filePath. Default value is FALSE.}
}
\value{
Either a plotly object (if interactive = TRUE) or a ggplot object
(if interactive = FALSE), save the output visualization plot to "filePath"
if save is TRUE
}
\description{
This is a function that can visualize the enrichment analysis
result as a Manhattan plot, whose x-axis contains different databases and
y-axis is "-Log10(P.adj)". Basically, the larger the value of -Log10(P.adj),
the more the pathway is enriched in genes from the query gene list.
}
\examples{
# Example 1:
# Using enrichment analysis output (enrichOutputList) available with the
# package
\dontrun{
enrichOutputList$gProfilerResult
enrichOutputList$enrichmentVis

# Plot a Manhattan plot for the enrichment analysis results
visEnrichAnaly(enrichOutputList)
}

}
\references{
Kolberg L, Raudvere U, Kuzmin I, Vilo J, Peterson H (2020). “gprofiler2-an R package for
gene list functional enrichment analysis and namespace conversion toolset g:Profiler.”
F1000Research, 9 (ELIXIR)(709). R package version 0.2.2.

R Core Team (2023). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/

Vaidyanathan R, Xie Y, Allaire J, Cheng J, Sievert C, Russell K (2023). htmlwidgets: HTML
Widgets for R. R package version 1.6.2,
https://CRAN.R-project.org/package=htmlwidgets

Wickham H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4
https://CRAN.R-project.org/package=ggplot2
}
