% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visEnrichAnalyLollipop}
\alias{visEnrichAnalyLollipop}
\title{Generate a Lollipop plot for the enrichment analysis result}
\usage{
visEnrichAnalyLollipop(enrichOutputList = NULL, filePath = NULL)
}
\arguments{
\item{enrichOutputList}{A named list containing the named list of enrichment
analysis results from gprofiler2 and a processed data frame based on the
results from gprofiler2 for visualization. The output of function
"enrichAnalysis" is a good example. Default value is NULL.}

\item{filePath}{A character string path to the directory that the user want
to store the visualization of the enrichment analysis result.
It is recommended to create a new directory to store the output .png file.
Default value is NULL. Should in the format: "/Path/to/the/directory/".}
}
\value{
A NULL, save the output visualization plot to "filePath"
}
\description{
This is a function that can visualize the enrichment analysis
result as a Lollipop plot, whose x-axis is the gene ratio and
y-axis is the name of the pathways. Basically, the larger the gene ratio is,
the more genes enrich in the pathway.
}
\examples{
# Example 1:

}
\references{
R Core Team (2023). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria.
\href{https://www.R-project.org/}{link}.

Wickham H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4
\href{ https://CRAN.R-project.org/package=ggplot2}{Link}.
}
