% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGEAnaly.R
\name{runGEAnaly}
\alias{runGEAnaly}
\title{Execute the whole gene expression pipeline in a single call of a function}
\usage{
runGEAnaly(geneCounts = NULL, sampleInfor = NULL, filePath = NULL)
}
\arguments{
\item{geneCounts}{A data frame containing the matrix of un-normalized read
counts of gene expression data, normally obtained from RNA-seq or other
sequencing experiments. The count number in the matrix should be
non-negative integers. The row names of the matrix should be the genes,
and the column names should be the sample IDs. Default value is NULL.}

\item{sampleInfor}{A data frame containing the comparison table between
sample IDs and sample group, e.g. control group or treated group.
The first column is the IDs or names of the samples, the second column
is the group that the sample belongs to. First column name shoud be
"sample", second column name should be "condition". Default value is NULL.}

\item{filePath}{A character string path to the directory that the user want
to store the output files. It is recommended to create a new directory.
Default value is NULL. Should in the format: "/Path/to/the/directory".}
}
\value{
A NULL, save the output files to "filePath"
}
\description{
This is a function that can execute the whole gene expression analysis
pipeline. It accept a gene counts matrix and sample information, and perform
differential expression analysis, after that, genes with significantly
different expression levels are selected and labelled. Then a volcano plot is
used for the visualization of the differential expression analysis. Subsequently,
those selected significant genes are sent for an enrichment analysis. Then
a Manhattan plot and a Lollipop plot are generated for the visualization of
enrichment analysis. All steps are run by default settings.
}
\examples{
# Example 1:
# Using gene counts matrix (geneCountsDiffExpression) and the sample
# information (sampleInforDiffExpression) available with the package
\dontrun{
dim(geneCountsDiffExpression) # 11000 rows, 126 columns
dim(sampleInforDiffExpression) # 126 rows, 2 columns

# Perform the whole gene expression analysis pipeline on the input gene
# counts matrix
runGEAnaly(geneCountsDiffExpression,
           sampleInforDiffExpression,
           filePath = getwd())
# You should see all visualizations and all results from each stage of the
# pipeline in the current working directory, which is the path from getwd()
}

}
