% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGEAnaly.R
\name{runGEAnalyCor}
\alias{runGEAnalyCor}
\title{Execute the gene correlation analysis pipeline in a single call of a function}
\usage{
runGEAnalyCor(geneCounts = NULL, filePath = NULL)
}
\arguments{
\item{geneCounts}{A data frame containing the matrix of gene expression data,
normally obtained from RNA-seq or other sequencing experiments.
The row names of the matrix should be the genes, and the column names
should be the sample IDs. Default value is NULL.}

\item{filePath}{A character string path to the directory that the user want
to store the output files. It is recommended to create a new directory.
Default value is NULL. Should in the format: "/Path/to/the/directory".}
}
\value{
A NULL, save the output files to "filePath"
}
\description{
This is a function that can perform gene correlation analysis and visualize
the result. It accepts a gene counts matrix, and perform gene correlation
analysis, after that, the result is visualized by a heatmap. All steps are
dun by default settings.
}
\examples{
# Example 1:
# Using gene counts matrix (geneCountsCorrelation) available with the package
\dontrun{
dim(geneCountsCorrelation) # 105 rows, 41 columns

# Perform the gene correlation analysis pipeline on the input gene
# counts matrix
runGEAnalyCor(geneCountsCorrelation,
              filePath = getwd())
# You should see visualizations and results from each stage of the
# pipeline in the current working directory, which is the path from getwd()
}

}
