% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visDeAnaly}
\alias{visDeAnaly}
\title{Generate a volcano plot for the differential expression analysis result}
\usage{
visDeAnaly(
  genes = NULL,
  logFCT = 2,
  padjT = 0.05,
  maxOverlap = 8,
  colorBlindF = FALSE,
  filePath = NULL,
  save = FALSE
)
}
\arguments{
\item{genes}{A data frame, which has a column called "group",
indicating if the gene is up expressed or down expressed or no change for
its expression. The output of function "labelGenes" is a good example.}

\item{logFCT}{A positive integer, the log fold change threshold used to
select genes with significant large differences in expression levels.
Default value is 2.}

\item{padjT}{A positive double between 0 and 1, used to select genes with
significant differences in expression levels. Default value is 0.05.}

\item{maxOverlap}{The max number of labels of genes in the figure that can
overlap. Exclude text labels that overlap too many things. Defaults to 8.}

\item{colorBlindF}{A boolean indicating whther to use colorblind-friendly
colors. Default is FALSE.}

\item{filePath}{A character string path to the directory that the user want
to store the visualization of the differential expression analysis result.
It is recommended to create a new directory to store the output .png file.
Default value is NULL. Should in the format: "/Path/to/the/directory".}

\item{save}{A boolean value to indicate if store the result into a file
at the file path. If set to TRUE, the output would be stored at the
filePath. Default value is FALSE.}
}
\value{
A ggplot object, which is the visualization and save the output
visualization plot to "filePath" if save is TRUE.
}
\description{
This is a function that can visualize the differential expression analysis
result as a volcano plot, whose x-axis is "Log2(Fold change)" and y-axis is
"-Log10(P.adj)". Basically, the larger the value of -Log10(P.adj) and the
absolute value of Log2(Fold change) are, the more significant the gene
expression differences between samples are.
}
\examples{
# Example 1:
# Using labelled genes (labelledGenes) available with the package
\dontrun{
dim(labelledGenes) # 18309 rows, 7 columns

# Plot a volcano plot for the differential expression analysis results
# with colorblind-friendly colors
volcanoplot <- visDeAnaly(labelledGenes,
                          colorBlindF = TRUE)
}

}
\references{
R Core Team (2023). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/

Wickham H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. https://CRAN.R-project.org/package=ggplot2
}
