% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{visEnrichAnalyLollipop}
\alias{visEnrichAnalyLollipop}
\title{Generate a Lollipop plot for the enrichment analysis result}
\usage{
visEnrichAnalyLollipop(enrichOutputList = NULL, filePath = NULL, save = FALSE)
}
\arguments{
\item{enrichOutputList}{A named list containing the named list of enrichment
analysis results from gprofiler2 and a processed data frame based on the
results from gprofiler2 for visualization. The output of function
"enrichAnalysis" is a good example. Default value is NULL.}

\item{filePath}{A character string path to the directory that the user want
to store the visualization of the enrichment analysis result.
It is recommended to create a new directory to store the output .png file.
Default value is NULL. Should in the format: "/Path/to/the/directory/".}

\item{save}{A boolean value to indicate if store the result into a file
at the file path. If set to TRUE, the output would be stored at the
filePath. Default value is FALSE.}
}
\value{
A ggplot object, and save the output visualization plot to "filePath"
}
\description{
This is a function that can visualize the enrichment analysis
result as a Lollipop plot, whose x-axis is the gene ratio and
y-axis is the name of the pathways. Basically, the larger the gene ratio is,
the more genes enrich in the pathway. The plot shows the top 10 pathways with
top10 highest gene ratio.
}
\examples{
# Example 1:
# Using enrichment analysis output (enrichOutputList) available with the
# package
\dontrun{
enrichOutputList$gProfilerResult
enrichOutputList$enrichmentVis

# Plot a Lollipop plot for the enrichment analysis results
visEnrichAnalyLollipop(enrichOutputList)

# You should see visualizations saved in the current working directory,
# which is the path from getwd()
}

}
\references{
R Core Team (2023). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/

Wickham H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4
https://CRAN.R-project.org/package=ggplot2
}
